Revert extract for legacy types
The `extract` doc for the legacy datetime types contained tokens that are not supported for these types. This was the case because the doc also included the `TimestampTz` type which is one of the ones. I removed this type, as it caused already some confusion about what can be done with the old types.